refactor(dsp-api): Align *RestService method signature and simplify *Handlers
