+++ webapi/src/main/scala/org/knora/webapi/slice/security/api/AuthenticationApiModule.scala
@@ -22,5 +22,6 @@ object AuthenticationApiModule { self =>
       AuthenticationEndpointsV2.layer,
       AuthenticationEndpointsV2Handler.layer,
       AuthenticationApiRoutes.layer,
+      AuthenticationRestService.layer,
     )
 }
